kireidjangoapp/requirements.txt
alabaster==0.7.13
asgiref==3.6.0
astroid==2.15.5
Babel==2.11.0
beautifulsoup4==4.11.2
black==23.1.0
blinker==1.6.2
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.0.1
click==8.1.3
crispy-bootstrap4==2022.1
cryptography==39.0.1
dash==2.9.3
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
defusedxml==0.7.1
dill==0.3.6
Django==3.2.18
django-allauth==0.52.0
django-bootstrap-datepicker-plus==5.0.3
django-countries==7.5.1
django-crispy-forms==2.0
django-extensions==3.2.1
django-humanize==0.1.2
django-iban==0.3.1
django-internationalflavor==0.4.3
django-invoicing==4.3.1
django-model-utils==4.3.1
django-money==3.1.0
django-payments==2.0.0
django-phonenumber-field==7.0.2
django-plotly-dash==2.1.4
django-shopping-cart==0.1
django-shortcuts==1.6
docutils==0.19
dpd-components==0.1.0
ERAlchemy==1.2.10
fepdf==0.1.9
Flask==2.3.1
fpdf==1.7.2
gitdb==4.0.10
GitPython==3.1.30
greenlet==2.0.2
httplib2==0.22.0
humanize==4.6.0
idna==3.4
imagesize==1.4.1
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
mercadopago==2.2.0
mypy-extensions==1.0.0
numpy==1.24.3
oauthlib==3.2.2
packaging==23.0
pandas==2.0.1
pathspec==0.11.0
pguard==0.4.0
phonenumberslite==8.13.7
Pillow==9.4.0
plantuml==0.3.0
platformdirs==3.0.0
plotly==5.14.1
psycopg2==2.9.6
py-moneyed==2.0
pycparser==2.21
pydantic==1.10.7
pydot==1.4.2
Pygments==2.14.0
pygraphviz==1.10
PyJWT==2.6.0
pylint==2.17.4
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==6.1.3
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.3.24
sqlparse==0.4.3
tenacity==8.2.2
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.14
Werkzeug==2.3.2
wrapt==1.15.0
